refactor(individuality): move the domain types to types.ts
