+++ product-sdk/packages/individuality/src/derive.ts
@@ -12,44 +12,10 @@
  * permanent, and `Caution` is a *projection* of the next absence rather than a
  * count of past ones. See the package skill for the plain-language version.
  */
-import type { AbsenceGracePolicy, PersonhoodState } from "./types.js";
+import type { PersonhoodInputs, PersonhoodParticipant, PersonhoodState } from "./types.js";
 
 /** Attendance history is one byte, so a window wider than 8 games is capped. */
 const HISTORY_BITS = 8;
-
-/**
- * A participant's game record, as read from `Score.Participants` and decoded.
- *
- * `attendanceHistory` is a rolling byte: bit 0 is the most recent game, `1`
- * means attended and `0` means absent.
- */
-export interface PersonhoodParticipant {
-    score: number;
-    streak: { tag: "Attended" | "Absent"; count: number };
-    attendanceHistory: number;
-    reachedPersonhood: boolean;
-    recognition: "ExternallyRecognized" | "NotRecognized" | "Suspended" | "Recognized";
-    lastAttendedGame: number | null;
-}
-
-/**
- * Everything the derivation needs, resolved for one account at one block.
- *
- * Distinct from `FinalizedSnapshot` in `types.ts`, which is the block
- * coordinates the read was pinned to. This is the chain state read *at* that
- * block.
- */
-export interface PersonhoodSnapshot {
-    isLitePerson: boolean;
-    participant: PersonhoodParticipant | null;
-    /**
-     * `Score.PersonhoodThreshold`. **This is a `u8` on chain**, but PAPI types
-     * both `u8` and `u32` as `number`, so a width mistake here typechecks and
-     * passes tests. Verified against the metadata blob on 2026-08-16.
-     */
-    personhoodThreshold: number;
-    policy: AbsenceGracePolicy;
-}
 
 /** Count set bits without allocating (Kernighan). */
 function countSetBits(value: number): number {
@@ -83,7 +49,7 @@ function projectedMisses(history: number, window: number): number {
  * Never throws: an inconsistent record degrades to `Suspended` rather than
  * failing the caller's render.
  */
-export function derivePersonhoodState(snapshot: PersonhoodSnapshot): PersonhoodState {
+export function derivePersonhoodState(snapshot: PersonhoodInputs): PersonhoodState {
     const { participant } = snapshot;
 
     // A participant record always wins over Lite personhood.
@@ -153,7 +119,7 @@ if (import.meta.vitest) {
     });
 
     /** Snapshot around a recognized member; override per test. */
-    const snapshot = (overrides: Partial<PersonhoodSnapshot> = {}): PersonhoodSnapshot => ({
+    const snapshot = (overrides: Partial<PersonhoodInputs> = {}): PersonhoodInputs => ({
         isLitePerson: false,
         participant: participant(),
         personhoodThreshold: 5,
